ex. Loiane

+++ src/com/loiane/cursojava/variavel/VariaveisInteiras.java
@@ -8,7 +8,8 @@ public class VariaveisInteiras {
         byte idade01 = 20;
         short idade02 = 21;
         int idade03 = 22;
-        int idade04 = 23;
+        long idade04 = 23; // se eu quiser deixar explicito que o valor q estou usando é um
+        //long eu posso usar um L qdo declarar o valor, exemplo long = 23l;
 
         System.out.println("valor variável idade01 " + idade01);
         System.out.println("valor variável idade02 " + idade02);
